requirements.txt
#ray
#tqdm
#hydra-core==1.2.0
#pytorch-lightning==1.8.6
setuptools==59.5.0
protobuf>=3.20.2,<4
rich==12.4.4
wandb==0.13.5
plotly
# matplotlib==3.4.3
timm
plotly
torchmetrics==0.10.2
scikit-learn
onnx
onnxscript
onnxruntime-gpu
